We are slowly getting rid of boost::program_options and replaced it for now with an apparently decent BSD-licensed, header-only library shipped with Marian.

We are still awaiting a couple of fixes from the public repo, but let's put this here already, so not merging yet.

[Frank:] Since the process is not complete in the public repo, I for now flag this as "Reject" to give myself peace of mind that it doesn't get merged unexpectedly. Once the discussion on public master is complete, then we can reactivate this and merge without much of a review.

Related work items: #90261

+++ src/data/corpus_sqlite.h
@@ -4,9 +4,6 @@
 #include <iostream>
 #include <random>
 
-#include <boost/algorithm/string.hpp>
-#include <boost/iterator/iterator_facade.hpp>
-
 #include "common/config.h"
 #include "common/definitions.h"
 #include "common/file_stream.h"
@@ -77,7 +74,7 @@ public:
         maxDims.resize(ex.size(), 0);
       for(size_t i = 0; i < ex.size(); ++i) {
         if(ex[i].size() > (size_t)maxDims[i])
-          maxDims[i] = ex[i].size();
+          maxDims[i] = (int)ex[i].size();
       }
       sentenceIds.push_back(ex.getId());
     }
